+++ src/main/java/com/bit/lot/flower/auth/system/admin/http/SystemAdminRestController.java
@@ -4,7 +4,6 @@ import bloomingblooms.response.CommonResponse;
 import com.bit.lot.flower.auth.common.valueobject.AuthId;
 import com.bit.lot.flower.auth.system.admin.dto.UpdateStoreManagerStatusDto;
 import com.bit.lot.flower.auth.system.admin.service.UpdateStoreMangerStatusService;
-import io.swagger.annotations.ApiOperation;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -29,17 +28,12 @@ public class SystemAdminRestController {
     return CommonResponse.success("업데이트 완료");
   }
 
-
-  @ApiOperation(value = "시스템 관리자 로그인", notes = "Authroization: Bearer 토큰 생성, Refresh토큰"
-      + "Redis에 생성, HttpOnlyCookie에 생성")
   @PostMapping("/system/admin/login")
   public ResponseEntity<String> login(){
     log.info("login is processed");
     return ResponseEntity.ok("시스템 관리자 로그인 완료");
   }
 
-  @ApiOperation(value = "시스템 관리자 로그아웃", notes = "Authroization: Bearer 토큰 제거, Refresh토큰"
-      + "Redis에서 제거, HttpOnlyCookie에서 제거")
   @PostMapping("/system/admin/logout")
   public ResponseEntity<String> logout() {
     return ResponseEntity.ok("시스템 관리자 로그아웃 완료");
